CHECKPOINT initial SuSE build upto `pkg/pillar`
Needed to swizzle packages, some are simple changes.  But the
`devel` is in the details -- that is, `dev` packages are often
`devel` in SuSE.  The granularity of the Alpine packages is
also often finer, and as such some times need to remove an
Alpine package name.  SuSE work bases itself on release
15 SP4.

Started `build-tools` using `registry.suse.com/bci/golang:1.18-16.43`
for the Docker file in `build-tools/src/scripts` this is based
on a language specific utility image provided by SuSE on the net.
Tried to get as close to the `go` version currently used by the
EVE Alping build-tools.

`pkg/alpine` and `pkg/alpine-base` provide the basic build image
for EVE.  The `base` image was used to bootstrap `pkg/alpine`.
The resulting cached container is inteneded to cache the set of
packages used in the process (that is within the cached container
the files reused until intentionally updated).  Our SuSE image
is based on `registry.suse.com/bci/bci-base:15.4`. This is a
relatively generic SuSE cloud image.  We create a base EVE container
root file system (`registry.suse.com/bci/bci-busybox:15.4`).  This
is one of the smallest images provided by SuSE for cloud and/or
SLEmicro.  This choice can be updated by modifying the `base_rootfs`
`FROM` declaration in `pkg/suse/Dockerfile`.

`pkg/suse`, `pkg/debug`, and `pkg/dom0-ztools` are rather a mess.
Often I commented out lines to keep a grip on what the original
intent was, and better understand the build operation.  Will try
to clear that out with time.  `pkg/suse` started life as the
`pkg/alpine` Dockerfile.  It was cut up by Jack the Zypper
(`zypper` being the SuSE package management tool - the pun intended
to match the type of work done on the Dockerfile).

`pkg/storage-init`, `pkg/fscrypt`, `pkg/dnsmasq`, `pkg/strongswan`,
and `pkg/gpt-tools` required modification.

There is one currently known issue in `pkg/pillar`.  The patching
process for `dhcpcd` -- which doesn't seem to exist in SuSE.
Patches are applied to what I believe are `dhcpcd` configuration
files (see https://zededa.atlassian.net/browse/EV-272 created to
track that issue).

Need to build the EVE image now, which may require porting other
Docker files.

Signed-off-by: Gerald (Bob) Lee <gerald@zededa.com>

SEE: EVE-253,EVE-254