CHECKPOINT -- pillar builds
Completed `pkg/pillar` build.  Need to provide the root file system
tar archive.  That is currently a binary "provided" (that is
dumped in by me).  Need to add an independent way to generate that
tar archive, or publish the root file system tar artifact where
it can be globally retrieved.

Also, comment out the patch of the dhcpcd configuration files.
May be as simple as adding the correct package, locating the
configuration files, or a general over-haul.

SEE: EV-258