CHKPT -- dnsmasq and strongswan build
Add required packages to `suse/mirros/main` and clean-up the
deploy script.  The `set $PKGS` and `[ $# -eq 0 ] || <install>`
attempts to install the build packages if PKGS is not set.
Tried to set it to both a signle space and two spaces in the
hopes that it would create a situation where `$#` would be `0`
and the execution of `set` with white spaces didn't spill the
current environment ... sigh!