CHKPNT -- add missing builds to `pkg/pillar`
Had to add `gpt-tools`, `dnsmasq`, and `strongswan` builds to the
converted build methods to successfully complete `pillar` for real.
This is phase one that builds all the packages that depend on `suse`
will complete the build -- generate the cache entries for the
linuxkit build containers, and add those hashes to the `pillar`
Dockerfile.