CHKPNT-more packages to the cache, eve-suse SHA
Process more container build directories in `pkg`.  Also, had to
modify `Dockerfile`s to keep it building.  Some minor changes to
the action parts of `Dockerfile` -- tried to keep changes to a
minimum.  Retained some of the comments along the way for clarity.

Changes to further scope out what makes sense in the build process
as some of the `zypper install`s seem not to persist into the
container builds -- pilot error of some kind?  Yet another time
to sync with the upstream, and settle the shared components.

Had to loop back and delele make_rootfs.sh from the Dockerfile
as well as deleting it in git.  Error messages suck, if linuxkit
is going to be so 'opinionated' it should be articulate enough
to help narrow down what it is whining about.

Also updates `pkg/suse` reference.  Change is on master,
but need to test/commit eveything to get it re-based...

The `alpine` Docker file refers to a specific cached image of itself.
Fixed the comment for the current stable SHA for this repo sub-dir.
Still need to understand it, and just how I get the SHA of something
That changes based on the SHA... sigh.  Probably a special docker
in docker in docker ... well dockers all the way down.

This should complete the exercise.  Will pick up the changes to
Makefile from the `linuxkit` fix.  The `.yml` change is purely
speculative, and we can save and re-add...

Signed-off-by: Gerald (Bob) Lee <gerald@zededa.com>

SEE: EVE-253,EVE-254