CHKPNT-- additions to package list, build changes
Added packages required to build more of the EVE run time containers.

Changes to further scope out what makes sense in the build process
as some of the `zypper install`s seem not to persist into the
container builds -- pilot error of some kind?  Yet another time
to sync with the upstream, and settle the shared components.