CHKPNT- fix self reference sha in comment
The `alpine` Docker file refers to a specific cached image of itself.
Fixed the comment for the current stable SHA for this repo sub-dir.
Still need to understand it, and just how I get the SHA of something
That changes based on the SHA... sigh.  Probably a special docker
in docker in docker ... well dockers all the way down.