CHKPNT add last changes needed
These changes allow building the `installer`.

Most of the work is in mkimage-raw-uefi, looked at using dracut
to build the ram disk image; however, would require rearrangement
of the target root file system.  To move along, did a git clone
of `mkinitfs`, and `lddtree`.  Had to add a patch so it would build
in the SuSE environment.

Some work with kvm-tools to get it to build.  Kind of a hack, the
shell we are using expands the compound expansion of CONF_(uname -m)
to a shell variable that can in turn be expanded to the machine
specific stuff.  Limited right now to the amd64 (x86_64).

Updating new eve-suse stable sub-tree hash ... yet once more.