CHKPNT fix `eve-suse-deploy.sh` PKGS handling
The input processing to add architecture specific packages, results
in the shell variables to be at least one space.  Adjusted the code
such that even an empty set of packages results in a single space
set in the variable.  Without, empty does a `set` with white spaces
as input, which dumps the environment variables.  Also added code
to print on standard error when the "machine" is an unsupported one.