CHKPNT add `hostname` for xen build
Fixes for `pkg/xen`, `pkg/mkconf`, `pkg/u-boot` and `pkg/mkimage-iso-efi`
Helps with build of said packages.  Also, identified a way to
invoke the helper funcions and get `amd64` only builds (at points
in the make file, evaluates to build all allowed platforms).
That is `LINUXKIT_BUILD_PLATFORMS_LIST=linux/amd64` on the command
line to narrow the focus to the `amd64` machine architecture.

Needed to add the hostname package to the package cache for the
`xen` work.

For `xen-tool` and `mkimage-iso-efi` build `mkinitfs` as it is not
preent in SuSE.  May still want to do it, SuSE uses `dracut` --
would require some re-arrangement.