CHKPNT build and run live image
To correct for problems found in the archive file generated by
`linuxkit`.  Fixed previous problem with UTF-8 paths in symlinks
by setting the LANG enviroment variable.  This compensates for
over-writing existing directory object with a symbolic link.
Creates a temp directory image of the archive without `dev`.
Contents of the existing directories are moved to the target of
the link (`var/lock` -> `run/lock`, and `etc/ssl/certs` ->
`var/lib/ca-certificates/pem`).  The new utility `tarfilter`
removes **all** of the entries from the archive as it is copied
into a new file.  That file is updated with the links, and the
newly moved contents.  The new corrected file replaces the
original archive.  The temp directory is then removed.

In `Makefile` backported MR 2996 adding getty to qemu hosted
EVE instances.

Tested in that the combination of `make ZARCH=amd64 live` and
`make ZARCH=amd64 run` provide a qemu instance running SL-EVE.