CHKPNT add dhcp-client deps, and LD_LIBRARY_PATH
The `pillar` executible `zedbox` failed to locate `libzfs.4.so`
and just failed.  The library was in `/usr/lib`.  Added an
export of LD_LIBRARY_PATH to `init.sh` in `pillar`.  The search
path includes /lib64, /usr/lib64, /lib, and /usr/lib.  Now
executes without failure.

`dhclinet` was not present in `pillar`, a missing dependency was
causing it not to install.  Added dependencies to the package
cache.  Now installs, and correctly configures `eth0` and `eth1`
(at least for qemu).

Cleaned up debug print in `pkg/storage-init/storage-init.sh`.
Some fixes for `pkg/gpt-toos/files/zboot` -- `lsbloock` works
slightly different in SuSE.  Compensate for that.