CHKPNT HACK force currentPartitioin to IMGA
For this POC, IMGA will always be the current partition from
boot.  Thus can force it for now.  Problem, comes on a
call into zboot.go from the zboot script detailed in the
gpt-tools package.  Validtes, but ends up with an empty
string for the name.  Lots of nested single and double
quotes.