Hacks to build "working" raw installer
Works, but has a number of problems.

`poweroff` won't work because it is in a `chroot`. Displays a
message to poweroff and pull the usbstick.  Requires manual
intervention.

The `lspci` in the `debug` container is the busybox version, so it
doesn't support `-D`.

`lsblk` doesn't provide partition labels, so needed to work
around using `blkid`. (similar issue with `findfs` in
pkg/storage-init/storage-init.sh).

Due to `busybox --install -s` not working, added required commands
during docker image creation. Also, to enable debugging, created
a local copy of the initiial `init` script used by the kernel to
find and mount the real root file system.