CHKPNT:xen-tools: curl in debug, xen-tools initrd
Added `curl` to the debug container, it is in the standard EVE,
but doesn't appear in SLEVE.

Need to customize the `initrd` build to match the SuSE environment.
`busybox` doesn't support `--install` option.  So need to bring
in the links as needed.  The SuSE busybox BCI does not cotain
`udhcpc`, so create the link in `/out` prior to the operation.
The SuSE base BCI (basis for the build container) does not use
`busybox`, so `/out` (normally the file sysetem used for operations)
us the rootfs image at `/out` to generate the `initrd`.

`udhcpc` action script interpreter string pointed to the Alpine
`busybox` location (`#!/bin/busybox sh`), modified it to the
SuSE location (`/usr/bin/busybox`).  Fixed scriptlet `qemu-ifup`
to use `ip` command not `brctl`.

Builds, boots, loads containers; however, EVE still cannot ping
the container.