pkg/guacd: Add the `guacd` EVE package to SLEVE
Standard EVE build downloads and builds the OSSP libuuid version
1.6.2.  In SuSE, we have a later conflict with `libfontconfig.so.1`,
which requires `uuid_copy@UUID_1.0`.  The `guacd` build requires
the OSSP version of `libuuid`; however, found a patch allowing
guacd to build with the `linux-util` version of `libuuid`.

Added the patch to the build, removed the local build of `libuuid`,
`libuuid` was already loaded, built the image, and booted EVE in
qemu.  Verified that the `guacd` OCI was present and the process
was running.