pkg/vtpm: build and add to EVE image
Added package, and adjustments to `Dockerfile` allowing `pkg/vtpm`
to build.  Changed the target for the infrastructural libraries
to `lib64` (SuSE usage, and didn't need to add `LD_LIBRARY_PATH` to
the shell code.

Issue with the version of `libprotoc`, SuSE ships an older version
that we use.  May need to update it.  Testing limited to it not
interfering, and running.

Built, booted, and identified `vtpm` in a qemu hosted EVE instance.