pkg/rhel: first updated/fixes
`linuxkit` requires everything committed to `git` to process it.
Thus had to make the initial commit.

Focus to get a single version of `go` across the build and the
application space.  Created a directory in the builder which
contains the architectures and go version in files.  Also,
includes tar file for both the builder architecture, and the
target architecture.  Tried to use `BUILDARCH`, but it is the
machine you are building on, not that of the build container
(may be emulation, oh say amd64).

Updated the `yum` commands, using the new and improved `dnf`.
`yum` was deprecated in RHEL 8, with symlinks for backward
compatibility.  RHEL 9 provides that backward compatibility,
but pressing ahead.  Setting the configuration file for `dnf`
up front to ignore weak dependencies and some other settings.
`eve-rhel-deploy.sh` only pull packages from the cache, to it
fails if any package is missing.  `dnf` will not download with
an install; it requires a `download` command to pull those
down.

Updated the file names in `pkg/rhel/mirrors` to reflect the
RHEL package sections (not certain of the names RH has for
these sections or groups if you will).  Added sufficient packages
for `pkg/fscrypt` until the missing `pam-devel` package is
strictly required.  Used `pkg/fscrypt` as it uses `go` to build,
and allowed for testing of the set-up and build of the
EVE package (will pass when we can get `pam-devel`).

Removed `pam-devel` from the `pkg/fscrypt` list of BUILD_PKGS,
did the build (builds a number of `go` files) that dies
when `pam-devel`'s missing header files are needed for the
compilation.  Give good assurance that the build container
is working correctly, and most work moving forward is enhancement
or adding packages to the list.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>