pkg/rhel: subscription login, cache fixes, files
Need to login to RH via the subscription manager for access to the
full set of packages (such as `pam-devel`).  Requires credentials
(username/password).  Login, build the cache, and then (unregister).
Dropping plan to try to separately download the packages and do
a add them to the cache from a local directory.  Moved some of
the commends around as I understand more and experiment.  No one
is doing quite what we are, so documentation usually points at
something else.

When we unregister, the cache remains, but all of the repo entries
in `/etc/yum.repos.d` are gone.  From a paper about building local
repos, created a set of cache only entries that never expires,
and only points to the cache.  In `eve-rhel-deploy.sh` we copy
the repo file from `/eve` special storage (the go tar file, the
files with the architectures inside, ...).  Save a `tgz` archive
of the package cache in `/eve`.  `build-cache.sh` intuits the
directory name, so we can detect rpms already installed (need a
different command to download, and must know where to put it --
default is the current working directory).  Add option to limit
dependencies (no weak dependencies).  Identify files to directly
download (not install with `--downloadonly`) by monitoring the
output of `dnf install`.  The deploy script got some cleanup for
`go` installation, and fixes for the run-time package installation
(that is `--installdir` is more of a `chroot`, need config,
metadata, and cache in its file tree).

Added more packages for EVE container builds.

Basic changes to `pkg/storage-init`, `pkg/debug`, and `pkg/fscrypt`.
Starts off the basic port work.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>