pkg/rhel: fixes for secret, go install, and plat
Lots of problems trying to code for multiple architectures.  The
rules for `ENV` and `ARG` make it hard to process.  Alterations
to select correct architecture string for filenames.  Better
detection of the `golang` package requests, and pull in the `go`
modules `golint` and `ineffasign` for `go-compile.sh`.  These
are only installed in the build instance, not the run-time.
Verified file path names matching correct architecture.

Working version of the secret code.  Can't fully commit it until
the `Makefile` side of the work is done.  Completed the work,
and tested using `docker buildx build` commands that pass through
the secret.

Added packages required to complete port of more EVE containers

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>