pkg/dom0-ztools: RELEVE port
Package lists and deploy script modified.  Had some standard
swizzling to do, such as only add `util-linux` in place of the
Alpine `util-linux-devel`.  Alpine provides `libcrypto` as a
separate package.  That version was 1.1, which requires a `compat`
package from RHEL to get that older version.  Getting the
correct version of `libcrypto` and align with or without
FIPS can be painful.

The candidate run-time rootfs has symlinks in its root.  This
can create problems doing automated copies.  (links go to the
current active root, not the expected destination).  Had to
compensate for the copy of zfs tools into that candidate rootfs,
and the final copy of the linuxkit additions to the rootfs into
the candidate runtime rootfs.

Built the container.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>