pkg/{guacd,vtpm}: enter STOPPED state on boot
Identified and characterized by Adnan.

Both have the same problem.  Libraries cannot be found.  For
`pkg/guacd`, they are in `/usr/lib`, but SuSE only does a default
search of `*/lib64`.  The run-time rootfs did not provide
`libssl.so.1.1`.  When the guacd libraries are moved they are
now placed in `/usr/lib64`..  When the guacd libraries are copied
they are now placed in `/usr/lib64`.

`pkg/vtpm` was missing `libz.so.1` (`libz.so.1.2.11`).  Added the
`libz1` package to the run-time rootfs package list.

Built and booted a live image to verify presence of and location
of libraries, and that the containers do not just stop. (identified
the need for libssl at this point.  Also, created a raw installer
and installed on hardware box.  Repeat validation steps.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>