Build and run RELEVE limited live image
Verifies the build process so far, and limited run time verification.

`Makefile` needed the `rhel` OCI added to the build (rather
than `alpine`.  Also, directly enumerate only the EVE containers
to include.  `docker-compose.yml` and `images/rootfs.yml.in` are
also commented to remove RELEVE images not strictly needed for
this exercise.  `mkimage-raw-efi` had a dangling SHELL directive
setting `ash` as the default shell,  Trying to run the image
fails as it cannot find `ash`.  It also looks for `sgdisk` in
`/usr/bin`, but RHEL puts it in `/usr/sgin`.  `mkrootfs-squash`
had problems with `bsdtar` not handling its input.  The arguments
to it precluded its use.  The input is not compressed, and `bsdtar`
ignores compression in this case.  Standard `tar` works for the
RHEL case, need to try it on SuSE.

The `make ... run` had a problem `invalid accelerator hvf`, logged
and extracted the qemu command, removed the offending option, and
ran the live image. Get a shell prompt, executes `ls` and `ps`.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>