pkg/kvm-tools: RELEVE port work
Standard set of changes, but had to add some packages for the
build to work (`bzip` and `diffutils`).

Left `libvncserver` out of the build with no ill effects noted.
Also, had to comment the `ldd` of tow locally built binaries.
SuSE reports "not found", RHEL throws a SEGFAULT error.  Appears
to be informational only, so just commented it out.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>