pkg/rhel: fixes add packages for build
Added packages required for xen-tools, kvm-tools, and apparmor.
Fixed a problem in the way the path to the cache is determined.
Needed to better insure dealing with a single line, and the
contents are well understood.  Some packages updated, and thus
needed to be added to the package cache list.

Set environment variable to inform `dnf` that working in a
container environment.

Fixed build-tools `Dockerfile`.  RHEL `uname -o` prints
`GNU/Linux` (operating system) which injects an extra layer
in the file name to download.  Used `uname -s` for kernel
name (Linux).  Now pulls the correct, if deprecated, go
utility.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>