Update container builds to latest pkg/rhel
Needed to get all the containers that build to use the same version
of `pkg/rhel`.  As part of that exercise, packages in the `glibc`
and `python3` chain of packages were updated upstream (at RHEL).
As such to successfully build, needed to add a package from one of
these two package chain.  Thus adding of `glibc-devel` or `python3`.

Locaated bug in `build-tools/src/scripts/Dockerfile`, needed to
chagne argument to `uname` to get just `Linux` not `GNU/Linux`.
Also now usses `dnf` not `yum` for package additions.

Built all the containers

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>