live image on-boarded by the controller
Pillar micro services all now configure and run.  System configures
and contracts the specified controller.  `lsblk` exhibits some of
the same issues that SuSE displayed.  Needed to replace it with
shell code and `blkid`.

The pillar Dockerfile required some unique copy changes because
of the RHEL root file system (symbolic links in the root that
are fully qualified, that is `/lib64` is a link to `/usr/lib64`.
During the build and copy operations, it can pull through the
link and get the wrong `/`.

Changes to `mkrootfs-squash` to handle the links, may want to
remove the `-h`.  Just a bit confusing as the only link in the
linuxkit rootfs is `/lib64`.

Built the live image, booted, and on-boarded on the controller.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>