pkg/xen-tools: RELEVE port (+pkg/eve)
May still need work on the initramfs portion.

Build liburing as the RHEL devel version of the package is in the
beta release (and not directly available -- we can point at it,
but won't know what other packages are updated in the beta repo).
Updated liburing version to the same version provided by RHEL.
A direct pull, so won't have RHEL specific patches.

Copied the liburing libraries from */lib to */lib64 so they can
be found by the RHEL build environment.

Built, booted, verified xen-tools running.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>