pkg/rhel:pin cloud image version add packages
RHEL Universal Base Image version moved forward, and broke our
build.  Pinned the version of the build base, and base root file
system images.  Pinned to close dates (about three months ago).
Also updaged build-tools.

Changes to the dependencies required adding packages, and/or
adding packages to individual container package list(s) (both
the build time, and deployment image).  Appears to be when a
newer version of the packages is provided, but dependency is
for a previous version.

Also, added the watchdog package to help with the installer.

Updating container/EVE package builds to use this version of the
package cache.  Also, restored `gpt-tools` use of lsblk.

This enables the build of the raw installer.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>