RHEL-EVE: changes for install and on-board
Changes required to build the raw installer, that in turn would
install on an x86_64 system, and on-board with a zcloud.

`pkg/gpt-tools/files/zboot` needed changes to impedence match
partition management scripts.

Changes required to build the raw installer, that in turn would
install on an x86_64 system, and on-board with a zcloud.

`pkg/gpt-tools/files/zboot` needed changes to impedance match
partition management scripts.

Makefile adjust disk image sizes, list of required packages, only
use `pkg/new-kernel`, and enumerate the EVE packages (containers)
being built (for now).

The files `docker-compose.yml` and `images/rootfs.yml.in modified
to support only the containers being used.

Modifications to files in `pkg/mkimage-raw-efi` needed to create
a raw installer that installs an x86_64 RHEL EVE image, which
in turn will on-board with zcloud.

Builds, installs, and on-boards. Will load and start container and
VM images.  However, those images cannot connect with the outside
world (for now).

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>