pkg/wlan: add wlan to RELEVE
Needed to add the wpa_supplicant to the package cache, and build
wireless_tools.  The wireless_tools package is available, but
through some obscure maintenance repository.  Also, added patch
for build flags modified for a dynamically linked version of the
tools.

Builds, runs, and appears to work correctly.  Needs the wireless
interface tested.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>