pkg/rhel: use hash for the package cache
Need to add the correct rhel credentials to get the correct linuxkit
signature for the package cache.  Also, pkg/eve/Dockerfile.in does
not include u-boot-tools (not found in RHEL, long search for
the source files, ignored for PoC).  Builds installer-raw, which
runs.

Missed pkg/vtpm in the list of zededa `pkgs` to build.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>